fix loadVisualProperties cases
In some cases, the attributes were being pulled from the object, instead of from the passed in collection.  This would cause some attributes to be missed if they were in style attributes.
Fix printf statements in svgpath printing
some code cleanup

+++ svg/xmlscan.h
@@ -68,7 +68,7 @@ namespace waavs {
     // It also handles numeric entities
     // It does not handle any other entities
     //
-    static int expandBasicEntities(const ByteSpan& inSpan, ByteSpan& outSpan)
+    static size_t expandBasicEntities(const ByteSpan& inSpan, ByteSpan& outSpan)
     {
         ByteSpan outCursor = outSpan;
